Adjusts M3StopWatch return types

+++ M3/M3StopWatch.h
@@ -5,7 +5,7 @@
 
 -(void)startWatch;
 -(uint64_t)nanoSeconds;
--(float)milliSeconds;
--(float)seconds;
+-(uint64_t)milliSeconds;
+-(double)seconds;
 
 @end
